+++ BlurCinnamon@klangman/files/BlurCinnamon@klangman/6.0/extension.js
@@ -33,6 +33,10 @@ const Mainloop      = imports.mainloop;
 const AppletManager = imports.ui.appletManager;
 const Lang          = imports.lang;
 
+// For PopupMenu effects
+const Applet        = imports.ui.applet;
+const PopupMenu     = imports.ui.popupMenu;
+
 const GaussianBlur = require("./gaussian_blur");
 
 const ANIMATION_TIME = 0.25;
@@ -43,9 +47,10 @@ let originalAnimateExpo;
 
 let settings;
 let blurPanels;
-let blurMainMenu;
+let blurPopupMenus;
 
-var blurExtensionThis;
+var blurPanelsThis;
+var blurPopupMenusThis;
 
 const BlurType = {
    None: 0,
@@ -77,17 +82,22 @@ function _animateVisibleOverview() {
    let radius = (settings.overviewOverride) ? settings.overviewRadius : settings.radius;
    let blendColor = (settings.overviewOverride) ? settings.overviewBlendColor : settings.blendColor;
    let opacity = (settings.overviewOverride) ? settings.overviewOpacity : settings.opacity;
+   let saturation = (settings.overviewOverride) ? settings.overviewSaturation : settings.saturation;
 
    // Get the overview's background image and add the BlurEffect to it if configured to do so
+   let desktopBackground = children[0];
    if (blurType > BlurType.None) {
       let fx;
-      let desktopBackground = children[0];
       if (blurType === BlurType.Simple) {
          fx =  new Clutter.BlurEffect();
       } else {
          fx = new GaussianBlur.GaussianBlurEffect( { radius: radius, brightness: 1, width: 0, height: 0 } );
       }
       desktopBackground.add_effect_with_name( "blur", fx );
+   }
+   if (saturation<100) {
+      let desat = new Clutter.DesaturateEffect({factor: (100-saturation)/100});
+      desktopBackground.add_effect_with_name( "desat", desat );
    }
    // Get the overview's backgroundShade child and set it's color to see-through solid black/"Color blend" color
    let backgroundShade = children[1];
@@ -111,9 +121,11 @@ function _animateVisibleExpo() {
    let radius = (settings.expoOverride) ? settings.expoRadius : settings.radius;
    let blendColor = (settings.expoOverride) ? settings.expoBlendColor : settings.blendColor;
    let opacity = (settings.expoOverride) ? settings.expoOpacity : settings.opacity;
+   let saturation = (settings.expoOverride) ? settings.expoSaturation : settings.saturation;
+
+   let desktopBackground = this._background
    if (blurType > BlurType.None) {
       let fx;
-      let desktopBackground = this._background
       if (blurType === BlurType.Simple) {
          fx =  new Clutter.BlurEffect();
       } else {
@@ -121,7 +133,10 @@ function _animateVisibleExpo() {
       }
       desktopBackground.add_effect_with_name( "blur", fx );
    }
-
+   if (saturation<100) {
+      let desat = new Clutter.DesaturateEffect({factor: (100-saturation)/100});
+      desktopBackground.add_effect_with_name( "desat", desat );
+   }
    // Create a shade, set it's color in accordance with the settings and make it invisible
    let backgroundShade = new St.Bin({style_class: 'workspace-overview-background-shade'});
    this._backgroundShade = backgroundShade;
@@ -135,6 +150,7 @@ function _animateVisibleExpo() {
       { opacity: Math.round(opacity*2.55), time: ANIMATION_TIME, transition: 'easeNone' } );
 }
 
+// This is an implementation of Panel._panelHasOpenMenus() that will be used in pre-Cinnamon 6.4 versions
 function panelHasOpenMenus() {
    return global.menuStackLength > 0;
 }
@@ -147,7 +163,7 @@ class BlurPanels {
       this._blurredPanels = [];
       this._blurExistingPanels();
 
-      blurExtensionThis = this; // Make the 'this' pointer available in patch functions
+      blurPanelsThis = this; // Make the 'this' pointer available in patch functions
 
       // Monkey patch panel functions so we can manage the blurred backgrounds when the panels are hidden/shown
       this._originalPanelEnable    = Panel.Panel.prototype.enable;
@@ -230,7 +246,7 @@ class BlurPanels {
    _blurPanel(panel, index) {
       let panelSettings = this._getPanelSettings(panel, index);
       if (!panelSettings ) return;
-      let [opacity, blendColor, blurType, radius] = panelSettings;
+      let [opacity, blendColor, blurType, radius, saturation] = panelSettings;
 
       let actor = panel.actor;
       let blurredPanel = panel.__blurredPanel;
@@ -276,6 +292,15 @@ class BlurPanels {
          blurredPanel.effect = fx;
          blurredPanel.background = background;
       }
+      if (saturation<100) {
+         if (!blurredPanel.background) {
+            let background = Meta.X11BackgroundActor.new_for_display(global.display);
+            global.overlay_group.add_actor(background);
+            blurredPanel.background = background;
+         }
+         let desat = new Clutter.DesaturateEffect({factor: (100-saturation)/100});
+         blurredPanel.background.add_effect_with_name( "desat", desat );
+      }
    }
 
    // This function will restore all panels to their original state and undo the monkey patching
@@ -318,6 +343,32 @@ class BlurPanels {
       }
    }
 
+   // An extension setting controlling the color saturation overlay was modified
+   updateSaturation() {
+      let panels = Main.getPanels();
+      for ( let i=0 ; i < this._blurredPanels.length ; i++ ) {
+         if (panels[i] && this._blurredPanels[i]) {
+            let panelSettings = this._getPanelSettings(panels[i], i);
+            if (panelSettings) {
+               let [opacity, blendColor, blurType, radius, saturation] = panelSettings;
+               let blurredPanel = panels[i].__blurredPanel;
+               if (!blurredPanel.background) {
+                  let background = Meta.X11BackgroundActor.new_for_display(global.display);
+                  global.overlay_group.add_actor(background);
+                  blurredPanel.background = background;
+               }
+               let effect = blurredPanel.background.get_effect("desat");
+               if (effect) {
+                  effect.set_factor((100-saturation)/100);
+               } else {
+                  let desat = new Clutter.DesaturateEffect({factor: (100-saturation)/100});
+                  blurredPanel.background.add_effect_with_name( "desat", desat );
+               }
+            }
+         }
+      }
+   }
+
    // An extension setting controlling how the dim overlay was modified
    updateColor() {
       let panels = Main.getPanels();
@@ -326,7 +377,7 @@ class BlurPanels {
             let panelSettings = this._getPanelSettings(panels[i], i);
             if (panelSettings) {
                let actor = panels[i].actor;
-               let [opacity, blendColor, blurType, radius] = panelSettings;
+               let [opacity, blendColor, blurType, radius, saturation] = panelSettings;
                if (settings.allowTransparentColorPanels) {
                   let [ret,color] = Clutter.Color.from_string( blendColor );
                   if (!ret) { [ret,color] = Clutter.Color.from_string( "rgba(0,0,0,0)" ); }
@@ -355,7 +406,7 @@ class BlurPanels {
          if (panels[i]) {
             let panelSettings = this._getPanelSettings(panels[i], i);
             if (panelSettings) {
-               let [opacity, blendColor, blurType, radius] = panelSettings;
+               let [opacity, blendColor, blurType, radius, saturation] = panelSettings;
                let blurredPanel = panels[i].__blurredPanel;
                if (blurredPanel) {
                   if (blurType !== BlurType.None && !blurredPanel.background) {
@@ -406,7 +457,7 @@ class BlurPanels {
                      continue;
                   }
                }
-               return [uniqueSetting.opacity, uniqueSetting.color, uniqueSetting.blurtype, uniqueSetting.radius];
+               return [uniqueSetting.opacity, uniqueSetting.color, uniqueSetting.blurtype, uniqueSetting.radius, uniqueSetting.saturation];
             }
          }
          return null;
@@ -415,7 +466,8 @@ class BlurPanels {
          let blurType = (settings.panelsOverride) ? settings.panelsBlurType : settings.blurType;
          let blendColor = (settings.panelsOverride) ? settings.panelsBlendColor : settings.blendColor;
          let opacity = (settings.panelsOverride) ? settings.panelsOpacity : settings.opacity;
-         return [opacity, blendColor, blurType, radius];
+         let saturation = (settings.panelsOverride) ? settings.panelsSaturation : settings.saturation;
+         return [opacity, blendColor, blurType, radius, saturation];
       }
    }
 
@@ -428,7 +480,7 @@ class BlurPanels {
                {opacity: 255, duration: AUTOHIDE_ANIMATION_TIME * 1000, mode: Clutter.AnimationMode.EASE_OUT_QUAD } );
          }
       } catch (e) {}
-      blurExtensionThis._originalPanelEnable.apply(this, params);
+      blurPanelsThis._originalPanelEnable.apply(this, params);
    }
 
    blurDisable(...params) {
@@ -439,7 +491,7 @@ class BlurPanels {
                   onComplete: () => { this.__blurredPanel.background.hide(); } });
          }
       } catch (e) {}
-      blurExtensionThis._originalPanelDisable.apply(this, params);
+      blurPanelsThis._originalPanelDisable.apply(this, params);
    }
 
    blurShowPanel(...params) {
@@ -452,216 +504,234 @@ class BlurPanels {
             } } );
          }
       } catch (e) {}
-      blurExtensionThis._originalPanelShowPanel.apply(this, params);
+      blurPanelsThis._originalPanelShowPanel.apply(this, params);
    }
 
    blurHidePanel(force) {
       try {
          let background = this.__blurredPanel.background;
          if (background && background.is_visible() && !this._destroyed && (!this._shouldShow || force) && !this._panelHasOpenMenus()) {
-            Tweener.addTween(background, {time: AUTOHIDE_ANIMATION_TIME, onUpdateScope: this, onUpdate: () => { 
+            Tweener.addTween(background, {time: AUTOHIDE_ANIMATION_TIME, onUpdateScope: this, onUpdate: () => {
                this.__blurredPanel.background.set_clip( this.actor.x, this.actor.y, this.actor.width, this.actor.height );
             }, onComplete: () => { background.hide(); } } );
          }
       } catch (e) {}
-      blurExtensionThis._originalPanelHidePanel.apply(this, force);
+      blurPanelsThis._originalPanelHidePanel.apply(this, force);
    }
 }
 
-// This class manages the blurring of the main menu(s)
-class BlurMainMenu {
-
+class BlurPopupMenus {
    constructor() {
-      this._timeWaited = 0;
-      this._signalManager = new SignalManager.SignalManager(null);
-      this.mainMenuApplets = [];
-      // We only need one background and only one blur effect since only one menu can be open at a time
-      this.background = null;
-      this.fx = null;
-      this._blurExistingMenus();
-      // Listen for new applets added to the panels
-      this._signalManager.connect(global.settings, "changed::enabled-applets", Lang.bind(this, this.onEnabledAppletsChanged) );
+      blurPopupMenusThis = this; // Make "this" available to monkey patched functions
+      this.original_popupmenu_open = PopupMenu.PopupMenu.prototype.open;
+      PopupMenu.PopupMenu.prototype.open = this._popupMenuOpen;
+      //this.original_popupmenu_close = PopupMenu.PopupMenu.prototype.close;
+      //PopupMenu.PopupMenu.prototype.close = this._popupMenuClose;
+
+      this._blurEffect = new GaussianBlur.GaussianBlurEffect( {radius: 0, brightness: 1 , width: 0, height: 0} );
+      this._desatEffect = new Clutter.DesaturateEffect({factor: 1});
+      this._background = Meta.X11BackgroundActor.new_for_display(global.display);
+      this._background.add_effect_with_name( "blur", this._blurEffect );
+      this._background.add_effect_with_name( "desat", this._desatEffect );
+      global.overlay_group.add_actor(this._background);
+      this._background.hide();
    }
 
-   // Apply the blur effects to all the existing menu applets
-   _blurExistingMenus() {
-      if (AppletManager.appletsLoaded) {
-         let applets = AppletManager.getRunningInstancesForUuid("menu@cinnamon.org");
-         for (let i=0 ; i < applets.length ; i++) {
-            this._blurMenu( applets[i] );
-         }
-      } else {
-         // The applects are not all loaded yes, so we need to wait a bit before we can be sure to see the main menu
-         if (this._timeWaited < 60) {
-            this._timeWaited++;
-            Mainloop.timeout_add( 1000, Lang.bind(this, this._blurExistingMenus) );
+   // Monkey patched over PupupMenu.open()
+   _popupMenuOpen(animate) {
+      if (this instanceof Applet.AppletPopupMenu || this instanceof Applet.AppletContextMenu || this instanceof Panel.PanelContextMenu) {
+         blurPopupMenusThis._blurPopupMenu(this);
+      }
+      blurPopupMenusThis.original_popupmenu_open.call(this, animate);
+   }
+
+   // Monkey patched over PopupMenu.close()
+   //_popupMenuClose(animate) {
+   //   blurPopupMenusThis.original_popupmenu_close.call(this, animate);
+   //}
+
+   // Set the visible section of the background based on the size of the popup menu
+   _setClip(menu){
+      if (menu && this._currentMenu && menu === this._currentMenu) {
+         let actor = menu.actor;
+         if (actor.visible) {
+            this._background.set_clip( actor.x, actor.y, actor.width, actor.height );
          } else {
-            log( `BlurCinnamon: Unable to find the Menu Menu after waiting for ${this._timeWaited} seconds for the applets to finish loading` );
+            this._background.set_clip( 0, 0, 0, 0 );
          }
       }
    }
 
-   // Apply the blur effects to the Menu Menu applet argument
-   _blurMenu(applet) {
-      let blurredMainMenu;
+   _onOpenStateChanged(menu, open) {
+      if (open) {
+         let radius = (settings.popupOverride) ? settings.popupRadius : settings.radius;
+         let blurType = (settings.popupOverride) ? settings.popupBlurType : settings.blurType;
+         let blendColor = (settings.popupOverride) ? settings.popupBlendColor : settings.blendColor;
+         let opacity = (settings.popupOverride) ? settings.popupOpacity : settings.opacity;
+         let accentOpacity = (settings.popupOverride) ? settings.popupAccentOpacity : Math.min(settings.opacity+10, 100);
+         let saturation = (settings.popupOverride) ? settings.popupSaturation : settings.saturation;
 
-      // Only effect the menu if the applet looks like we expect (protect against breaking things in case the menu applet changes in future releases)
-      if (applet && applet.menu && applet.menu.box && applet.searchEntry && applet.left_box && applet._resizer && applet._resizer._resized_callback) {
-         let box = applet.menu.box;
-         let entry = applet.searchEntry;
-         let leftBox = applet.left_box;
-         this.mainMenuApplets.push( { applet: applet,
-            original_box_color: box.get_background_color(), original_box_style: box.get_style(), original_box_class: box.get_style_class_name(), original_box_pseudo_class: box.get_style_pseudo_class(),
-            original_entry_color: entry.get_background_color(), original_entry_style: entry.get_style(), original_entry_class: entry.get_style_class_name(), original_entry_pseudo_class: box.get_style_pseudo_class(),
-            original_leftBox_color: leftBox.get_background_color(), original_leftBox_style: leftBox.get_style(), original_leftBox_class: leftBox.get_style_class_name(), original_leftBox_pseudo_class: leftBox.get_style_pseudo_class()
-            } );
-         if (settings.allowTransparentColorMainMenu) {
+         if (settings.allowTransparentColorPopup) {
+            let box = menu.box;
+            menu.blurCinnamonData.push( {entry: box, original_entry_color: box.get_background_color(), original_entry_style: box.get_style(),
+                                 original_entry_class: box.get_style_class_name(), original_entry_pseudo_class: box.get_style_pseudo_class()} );
             box.set_style( "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
                            "background-gradient-direction: vertical; background-gradient-start: transparent; " +
                            "background-gradient-end: transparent;    background: transparent;" );
-            entry.set_style( "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
-                             "background-gradient-direction: vertical; background-gradient-start: transparent; " +
-                             "background-gradient-end: transparent;    background: transparent;" );
-            leftBox.set_style( "background-gradient-direction: vertical; background-gradient-start: transparent; " +
-                               "background-gradient-end: transparent;    background: transparent;" );
-         }
-         this._signalManager.connect(applet.menu, "open-state-changed", Lang.bind(this, this._onOpenStateChanged) );
-      }
-   }
-
-   // Handle a resize event for the currently open menu
-   _onResize(w, h) {
-      this.origCallback.apply(this.openMenu, [w, h]);
-      if (this.openMenu) {
-         let actor = this.openMenu.actor;
-         let margin = actor.get_margin();
-         this.background.set_clip( actor.x+margin.left, actor.y+margin.top, actor.width-(margin.left+margin.right), actor.height-(margin.top+margin.bottom) );
-      }
-   }
-
-   // Handle the event of a Menu Menu opening or closing
-   _onOpenStateChanged(menu, open) {
-      if (open) {
-         let margin = menu.actor.get_margin();
-         let radius = (settings.mainmenuOverride) ? settings.mainmenuRadius : settings.radius;
-         let blurType = (settings.mainmenuOverride) ? settings.mainmenuBlurType : settings.blurType;
-         let blendColor = (settings.mainmenuOverride) ? settings.mainmenuBlendColor : settings.blendColor;
-         let opacity = (settings.mainmenuOverride) ? settings.mainmenuOpacity : settings.opacity;
-         let accentOpacity = (settings.mainmenuOverride) ? settings.mainmenuAccentOpacity : Math.min(settings.opacity+10, 100);
-         if (!this.background) {
-            this.background = Meta.X11BackgroundActor.new_for_display(global.display);
-            global.overlay_group.add_actor(this.background);
-            if (blurType === BlurType.Simple) {
-               this.fx =  new Clutter.BlurEffect();
-            } else {
-               this.fx = new GaussianBlur.GaussianBlurEffect( {radius: radius, brightness: 1 , width: 0, height: 0} );
-            }
-            this.background.add_effect_with_name( "blur", this.fx );
-         } else {
-            if (blurType === BlurType.Simple && this.fx instanceof GaussianBlur.GaussianBlurEffect) {
-               this.background.remove_effect(this.fx);
-               this.fx =  new Clutter.BlurEffect();
-               this.background.add_effect_with_name( "blur", this.fx );
-            } else if (blurType === BlurType.Gaussian && this.fx instanceof Clutter.BlurEffect) {
-               this.background.remove_effect(this.fx);
-               this.fx = new GaussianBlur.GaussianBlurEffect( {radius: radius, brightness: 1, width: 0, height: 0} );
-               this.background.add_effect_with_name( "blur", this.fx );
-            } else if (blurType === BlurType.Gaussian && this.fx.radius !== radius) {
-               this.fx.radius = radius;
-            }
-            if (blurType !== BlurType.None) {
-               this.background.show();
-            }
-         }
-
-         if (settings.allowTransparentColorMainMenu) {
-            // The menu's class seems to be updated every time the menu is opened, so we have to reset it every time to be transparent again!
-            // Since it's reset every time anyhow, we don't need to remember the style and restore it when/if menu effects are disabled
+            // Since menu.actor style is reset every time anyhow, we don't need to remember the style and restore it when the menu closes
             menu.actor.set_style(  "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
                            "background-gradient-direction: vertical; background-gradient-start: transparent; " +
                            "background-gradient-end: transparent;    background: transparent;"  );
+
+            // Set the popup menu color
+            let [ret,color] = Clutter.Color.from_string( blendColor );
+            if (!ret) { [ret,color] = Clutter.Color.from_string( "rgba(0,0,0,0)" ); }
+            color.alpha = opacity*2.55;
+            menu.box.set_background_color(color);
+
+            // Setup the popup menu accent actor color
+            color.alpha = (accentOpacity)*2.55;
+            this._accentColor = color;
+            // Find all the accent actors and adjust their transparency and background color
+            this._findAccentActors(menu, menu.actor);
+         }
+         // Setup the blur effect properly
+         let curEffect = this._background.get_effect("blur");
+         if (blurType === BlurType.None && curEffect) {
+            this._background.remove_effect(this._blurEffect);
+         } else if (blurType === BlurType.Simple && !(this._blurEffect instanceof Clutter.BlurEffect)) {
+            if (curEffect) {
+               this._background.remove_effect(this._blurEffect);
+            }
+            this._blurEffect =  new Clutter.BlurEffect();
+            this._background.add_effect_with_name( "blur", this._blurEffect );
+         } else if (blurType === BlurType.Gaussian && !(this._blurEffect instanceof GaussianBlur.GaussianBlurEffect)) {
+            if (curEffect) {
+               this._background.remove_effect(this._blurEffect);
+            }
+            this._blurEffect = new GaussianBlur.GaussianBlurEffect( {radius: radius, brightness: 1, width: 0, height: 0} );
+            this._background.add_effect_with_name( "blur", this._blurEffect );
+         } else if (blurType !== BlurType.None && curEffect === null) {
+            this._background.add_effect_with_name( "blur", this._blurEffect );
+         }
+         // Adjust the effects
+         if (this._blurEffect instanceof GaussianBlur.GaussianBlurEffect && this._blurEffect.radius != radius) {
+            this._blurEffect.radius = radius;
+         }
+         if (this._desatEffect.factor !== (100-saturation)/100) {
+            this._desatEffect.set_factor((100-saturation)/100);
          }
 
-         // Set the menu color
-         let [ret,color] = Clutter.Color.from_string( blendColor );
-         if (!ret) { [ret,color] = Clutter.Color.from_string( "rgba(0,0,0,0)" ); }
-         color.alpha = opacity*2.55;
-         menu.box.set_background_color(color);
-
-         color.alpha = (accentOpacity)*2.55;
-         menu.launcher.searchEntry.set_background_color(color);
-         menu.launcher.left_box.set_background_color(color);
-
-         this.background.set_clip( menu.actor.x+margin.left, menu.actor.y+margin.top, menu.actor.width-(margin.left+margin.right), menu.actor.height-(margin.top+margin.bottom) );
-         // Monkey patch the resize callback so we can intercept the resize events
-         this.origCallback = menu.launcher._resizer._resized_callback;
-         menu.launcher._resizer._resized_callback = Lang.bind( this, this._onResize );
-         this.openMenu = menu;
-      } else {
-         if (this.background) {
-           this.background.hide();
+         this._currentMenu = menu;
+         if (menu.animating) {
+            // Make the background visible but zero size initially, let the paint signal re-clip the background as needed
+            this._background.set_clip( 0, 0, 0, 0 );
+         } else {
+            let actor = menu.actor;
+            let margin = actor.get_margin();
+            this._background.set_clip( actor.x+margin.left, actor.y+margin.top, actor.width-(margin.left+margin.right), actor.height-(margin.top+margin.bottom) );
          }
-         // Restore the monkey patched callback
-         menu.launcher._resizer._resized_callback = this.origCallback;
-         this.openMenu = null;
+         this._background.show();
+         // Now that the menu is open we need to know if new actors are added so we can check for accent elements
+         menu.blurCinnamonSignalManager.connect(menu.actor, "queue-relayout", () => {this._findAccentActors(menu, menu.actor);} );
       }
    }
 
-   // Handle the event of a new Applet being added to a panel, look for new menu applets
-   onEnabledAppletsChanged() {
-      let i;
-      let applets = AppletManager.getRunningInstancesForUuid("menu@cinnamon.org");
-      for (i=0 ; i < applets.length ; i++) {
-         if (!this.mainMenuApplets.find( (mmData) => mmData.applet === applets[i] )) {
-            this._blurMenu( applets[i] );
-         }
-      }
-      /* // There is really noting to do when a menu menu applet is removed I think
-      log( "Looking for removed applets" );
-      for (i=0 ; i < this.mainMenuApplets ; i++) {
-         if (!applets.find( (applet) => applet === this.mainMenuApplets[i].applet )) {
-            log( "Removing main menu" );
-         }
-      }*/
+   // Called when Popup method is now closed and the animation is complete!
+   _onClosed(menu) {
+      this._unblurPopupMenu(menu);
    }
 
-   // Remove all the Main Menu effects and clean up (extension disabled or Main Menu effects are disabled)
+   // Look for Popup menu accent actors
+   _findAccentActors(menu, actor, indent="") {
+      let children = actor.get_children();
+      for (let i=0 ; i < children.length ; i++ ) {
+         let child = children[i];
+         if (child instanceof St.Entry) {
+            if (!menu.blurCinnamonData.find( (element) => {return(element.entry === child);} )) {
+               menu.blurCinnamonData.push( {entry: child, original_entry_color: child.get_background_color(), original_entry_style: child.get_style(),
+                                original_entry_class: child.get_style_class_name(), original_entry_pseudo_class: child.get_style_pseudo_class()} );
+            }
+            child.set_style( "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
+                             "background-gradient-direction: vertical; background-gradient-start: transparent; " +
+                             "background-gradient-end: transparent;    background: transparent;" );
+            child.set_background_color(this._accentColor);
+         } else if (child instanceof St.BoxLayout) {
+            let styleClassName = child.get_style_class_name();
+            if (styleClassName && styleClassName.indexOf("menu-favorites-box") !== -1) {
+               // This is for the menu@cinnamon.org applet
+               if (!menu.blurCinnamonData.find( (element) => {return(element.entry === child);} )) {
+                  menu.blurCinnamonData.push( {entry: child, original_entry_color: child.get_background_color(), original_entry_style: child.get_style(),
+                                original_entry_class: child.get_style_class_name(), original_entry_pseudo_class: child.get_style_pseudo_class()} );
+               }
+               child.set_style( "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
+                                "background-gradient-direction: vertical; background-gradient-start: transparent; " +
+                                "background-gradient-end: transparent;    background: transparent;" );
+               child.set_background_color(this._accentColor);
+            }
+         } else if (child instanceof St.Table) {
+            let name = child.get_name();
+            if (name && name.indexOf("notification") !== -1) {
+               // Notification messages in the Notifications applet, these messages can be removed by the user so we need to handle that case
+               if (!menu.blurCinnamonData.find( (element) => {return(element.entry === child);} )) {
+                  menu.blurCinnamonData.push( {entry: child, original_entry_color: child.get_background_color(), original_entry_style: child.get_style(),
+                                original_entry_class: child.get_style_class_name(), original_entry_pseudo_class: child.get_style_pseudo_class()} );
+                  menu.blurCinnamonSignalManager.connect(child, 'destroy', (actor) =>
+                     {
+                        // If the child is removed then remove the restore data to avoid errors
+                        let idx = menu.blurCinnamonData.findIndex( (element) => {return(element.entry === child);} )
+                        if (idx!=-1) {
+                           menu.blurCinnamonData.splice(idx,1);
+                        }
+                     } );
+               }
+               child.set_style( "border-radius: 0px; " + //"border-image: none;  border-color: transparent;  box-shadow: 0 0 transparent; " +
+                                "background-gradient-direction: vertical; background-gradient-start: transparent; " +
+                                "background-gradient-end: transparent;    background: transparent;" );
+               child.set_background_color(this._accentColor);
+            }
+         }
+         //if (child instanceof St.Widget) {
+         //   log( `Widget: ${indent}${child}` );
+         //}
+         this._findAccentActors(menu, child, indent+"  ");
+      }
+   }
+
+   _blurPopupMenu(menu) {
+      menu.blurCinnamonData = [];
+      menu.blurCinnamonSignalManager = new SignalManager.SignalManager(null);
+      menu.blurCinnamonSignalManager.connect(menu, "open-state-changed", Lang.bind(this, this._onOpenStateChanged) );
+      menu.blurCinnamonSignalManager.connect(menu, "menu-animated-closed", Lang.bind(this, this._onClosed) );
+      menu.blurCinnamonSignalManager.connect(menu.actor, 'notify::size', () => {this._setClip(menu);} );
+      menu.blurCinnamonSignalManager.connect(menu.actor, 'notify::position', () => {this._setClip(menu);} );
+   }
+
+   _unblurPopupMenu(menu) {
+      menu.blurCinnamonSignalManager.disconnectAllSignals();
+      delete menu.blurCinnamonSignalManager;
+      let box = menu.box;
+      // Restore the menu to it's original state
+      for (let i=0 ; i < menu.blurCinnamonData.length ; i++) {
+         let entry = menu.blurCinnamonData[i];
+         entry.entry.set_background_color(entry.original_entry_color);
+         entry.entry.set_style(entry.original_entry_style);
+         entry.entry.set_style_class_name(entry.original_entry_class);
+         entry.entry.set_style_pseudo_class(entry.original_entry_pseudo_class);
+      }
+      delete menu.blurCinnamonData;
+      if (this._currentMenu === menu) {
+         this._background.hide();
+         this._currentMenu = null;
+      }
+   }
+
    destroy() {
-      this._signalManager.disconnectAllSignals();
-
-      for (let i=0 ; i < this.mainMenuApplets.length ; i++) {
-         let mmData = this.mainMenuApplets[i];
-         let applet = mmData.applet;
-         let box = applet.menu.box;
-         let entry = applet.searchEntry;
-         let leftBox = applet.left_box;
-
-         box.set_background_color(mmData.original_box_color);
-         box.set_style(mmData.original_box_style);
-         box.set_style_class_name(mmData.original_box_class);
-         box.set_style_pseudo_class(mmData.original_box_pseudo_class);
-
-         entry.set_background_color(mmData.original_entry_color);
-         entry.set_style(mmData.original_entry_style);
-         entry.set_style_class_name(mmData.original_entry_class);
-         entry.set_style_pseudo_class(mmData.original_entry_pseudo_class);
-
-         leftBox.set_background_color(mmData.original_leftBox_color);
-         leftBox.set_style(mmData.original_leftBox_style);
-         leftBox.set_style_class_name(mmData.original_leftBox_class);
-         leftBox.set_style_pseudo_class(mmData.original_leftBox_pseudo_class);
-      }
-
-      if (this.background) {
-         this.background.remove_effect(this.fx);
-         global.overlay_group.remove_actor(this.background);
-         this.background.destroy();
-         this.background = null;
-         this.fx = null;
-      }
+      // Restore monkey patched PopupMenu open & close functions
+      PopupMenu.PopupMenu.prototype.open = this.original_popupmenu_open;
+      //PopupMenu.PopupMenu.prototype.close = this.original_popupmenu_close;
    }
 }
-
 
 class BlurSettings {
    constructor(uuid) {
@@ -670,42 +740,53 @@ class BlurSettings {
       this.settings.bind('blurType',   'blurType',   blurChanged);
       this.settings.bind('radius',     'radius',     blurChanged);
       this.settings.bind('blendColor', 'blendColor', colorChanged);
+      this.settings.bind('saturation', 'saturation', saturationChanged);
 
       this.settings.bind('overview-opacity',    'overviewOpacity');
       this.settings.bind('overview-blurType',   'overviewBlurType');
       this.settings.bind('overview-radius',     'overviewRadius');
       this.settings.bind('overview-blendColor', 'overviewBlendColor');
+      this.settings.bind('overview-saturation', 'overviewSaturation');
 
       this.settings.bind('expo-opacity',    'expoOpacity');
       this.settings.bind('expo-blurType',   'expoBlurType');
       this.settings.bind('expo-radius',     'expoRadius');
       this.settings.bind('expo-blendColor', 'expoBlendColor');
+      this.settings.bind('expo-saturation', 'expoSaturation');
 
       this.settings.bind('panels-opacity',    'panelsOpacity',    colorChanged);
       this.settings.bind('panels-blurType',   'panelsBlurType',   blurChanged);
       this.settings.bind('panels-radius',     'panelsRadius',     blurChanged);
       this.settings.bind('panels-blendColor', 'panelsBlendColor', colorChanged);
+      this.settings.bind('panels-saturation', 'panelsSaturation', saturationChanged);
 
-      this.settings.bind('mainmenu-opacity',        'mainmenuOpacity');
-      this.settings.bind('mainmenu-accent-opacity', 'mainmenuAccentOpacity');
-      this.settings.bind('mainmenu-blurType',       'mainmenuBlurType');
-      this.settings.bind('mainmenu-radius',         'mainmenuRadius');
-      this.settings.bind('mainmenu-blendColor',     'mainmenuBlendColor');
-      this.settings.bind('allow-transparent-color-mainmenu', 'allowTransparentColorMainMenu', mainMenuChanged);
+      this.settings.bind('popup-opacity',        'popupOpacity');
+      this.settings.bind('popup-accent-opacity', 'popupAccentOpacity');
+      this.settings.bind('popup-blurType',       'popupBlurType');
+      this.settings.bind('popup-radius',         'popupRadius');
+      this.settings.bind('popup-blendColor',     'popupBlendColor');
+      this.settings.bind('popup-saturation',     'popupSaturation');
+      this.settings.bind('allow-transparent-color-popup', 'allowTransparentColorPopup');
 
       this.settings.bind('enable-overview-override', 'overviewOverride');
       this.settings.bind('enable-expo-override',     'expoOverride');
       this.settings.bind('enable-panels-override',   'panelsOverride', panelsSettingsChangled);
-      this.settings.bind('enable-mainmenu-override', 'mainmenuOverride');
+      this.settings.bind('enable-popup-override',    'popupOverride');
 
       this.settings.bind('enable-overview-effects', 'enableOverviewEffects', enableOverviewChanged);
       this.settings.bind('enable-expo-effects',     'enableExpoEffects',     enableExpoChanged);
       this.settings.bind('enable-panels-effects',   'enablePanelsEffects',   enablePanelsChanged);
-      this.settings.bind('enable-mainmenu-effects', 'enableMainMenuEffects', enableMainMenuChanged);
+      this.settings.bind('enable-popup-effects', 'enablePopupEffects', enablePopupChanged);
 
       this.settings.bind('enable-panel-unique-settings', 'enablePanelUniqueSettings');
       this.settings.bind('panel-unique-settings', 'panelUniqueSettings', panelsSettingsChangled);
       this.settings.bind('allow-transparent-color-panels', 'allowTransparentColorPanels', colorChanged);
+   }
+}
+
+function saturationChanged() {
+   if (blurPanels) {
+      blurPanels.updateSaturation();
    }
 }
 
@@ -725,6 +806,7 @@ function panelsSettingsChangled() {
    if (blurPanels) {
       blurPanels.updateBlur();
       blurPanels.updateColor();
+      blurPanels.updateSaturation();
    }
 }
 
@@ -757,19 +839,12 @@ function enablePanelsChanged() {
    }
 }
 
-function enableMainMenuChanged() {
-   if (blurMainMenu && !settings.enableMainMenuEffects) {
-      blurMainMenu.destroy();
-      blurMainMenu = null;
-   } else if (!blurMainMenu && settings.enableMainMenuEffects ) {
-      blurMainMenu = new BlurMainMenu();
-   }
-}
-
-function mainMenuChanged() {
-   if (blurMainMenu) {
-      blurMainMenu.destroy();
-      blurMainMenu = new BlurMainMenu();
+function enablePopupChanged() {
+   if (blurPopupMenus && !settings.enablePopupEffects) {
+      blurPopupMenus.destroy();
+      blurPopupMenus = null;
+   } else if (!blurPopupMenus && settings.enablePopupEffects ) {
+      blurPopupMenus = new BlurPopupMenus();
    }
 }
 
@@ -797,9 +872,9 @@ function enable() {
    if (settings.enablePanelsEffects) {
       blurPanels = new BlurPanels();
    }
-   // Create a Main Menu Effects class instance, the constructor will kick things off
-   if (settings.enableMainMenuEffects) {
-      blurMainMenu = new BlurMainMenu();
+   // Create a Popup menu Effects class instance, the constructor will set everything up.
+   if (settings.enablePopupEffects) {
+      blurPopupMenus = new BlurPopupMenus();
    }
 }
 
@@ -819,8 +894,8 @@ function disable() {
       blurPanels = null;
    }
 
-   if (blurMainMenu) {
-      blurMainMenu.destroy();
-      blurMainMenu = null;
+   if (blurPopupMenus) {
+      blurPopupMenus.destroy();
+      blurPopupMenus = null;
    }
 }